Additional cleanup and logic for P3
* Separates main.py god class into separate files
* Adds logic to detect collisions that occur from stopping train(s)
* Corrects logic for avoiding collisions due to failed trains
