--- main/objects.py
@@ -1,2 +1,0 @@
-class Train:
-    pass
